requirements.txt
alembic==1.10.3
click==8.1.3
colorama==0.4.6
Flask==2.2.3
greenlet==2.0.2
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.6
python-dotenv==1.0.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
